Decompile safari_zone

+++ include/pm_version.h
@@ -1,7 +1,7 @@
 #ifndef POKEHEARTGOLD_PM_VERSION_H
 #define POKEHEARTGOLD_PM_VERSION_H
 
-extern const int gGameVersion;
-extern const int gGameLanguage;
+extern const u8 gGameVersion;
+extern const u8 gGameLanguage;
 
 #endif //POKEHEARTGOLD_PM_VERSION_H
